chore(entities): rename node for javascript as it also works in browser
